@@ -13,21 +13,19 @@ public class AstahConnectionBuilder {
 
     private final Map<String, AstahNode> nodes;
 
-    private final AstahTheme theme;
-
     private final List<ConnectionBuilder> builders;
 
     public AstahConnectionBuilder(Map<String, AstahNode> nodes, AstahTheme theme) {
         this.nodes = nodes;
-        this.theme = theme;
         builders = Arrays.asList(
+                new AttributeConnectionBuilder(nodes, theme),
                 new DependencyConnectionBuilder(nodes, theme),
                 new UsageConnectionBuilder(nodes, theme),
                 new DiagramConnectionBuilder(nodes, theme),
                 new GeneralizationConnectionBuilder(nodes, theme),
                 new RealizationConnectionBuilder(nodes, theme),
                 new IncludeConnectionBuilder(nodes, theme)
-        );
+                );
     }
 
     public Map<String, AstahConnection> build() {
