@@ -3,6 +3,8 @@ package hu.modeldriven.astah.traceability.layout.impl.render;
 import hu.modeldriven.astah.traceability.layout.ConnectionRenderer;
 import hu.modeldriven.astah.traceability.layout.Path;
 import hu.modeldriven.astah.traceability.layout.impl.AstahConnection;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.awt.*;
 import java.awt.geom.Point2D;
@@ -10,15 +12,13 @@ import java.util.List;
 
 public class AstahConnectionRenderer implements ConnectionRenderer {
 
-    private final static int ICON_LABEL_PADDING = 5;
-
-    private final static int ICON_COORDINATE_FIX = 1;
-
+    private static final int ICON_LABEL_PADDING = 5;
+    private static final int ICON_COORDINATE_FIX = 1;
+    private final Logger logger = LoggerFactory.getLogger(AstahConnectionRenderer.class);
     private final AstahConnection connection;
     private final String name;
 
     private final AstahTheme theme;
-
     private final Image image;
 
     public AstahConnectionRenderer(AstahConnection connection, AstahTheme theme) {
@@ -30,6 +30,13 @@ public class AstahConnectionRenderer implements ConnectionRenderer {
 
     @Override
     public void render(Graphics2D g, Path path) {
+
+        logger.info("Trying to render connection: {}", name);
+
+        if (path.coordinates().size() < 2) {
+            logger.error("Path did not have at least two coordinates, not painting");
+            return;
+        }
 
         Color edgeColor;
         Color labelColor;
@@ -52,6 +59,8 @@ public class AstahConnectionRenderer implements ConnectionRenderer {
 
         int posX = (int) labelPosition.getX();
         int posY = (int) labelPosition.getY();
+
+        logger.info("Label positions for {}: {} {}", name,  posX, posY);
 
         // Because labels are not drawn by swing to the top left coordinate, this
         // has to be fixed with ascent calculation
