@@ -1,9 +1,6 @@
-package hu.modeldriven.astah.traceability;
+package hu.modeldriven.astah.traceability.plugin;
 
 
-import com.change_vision.jude.api.inf.AstahAPI;
-import com.change_vision.jude.api.inf.exception.ProjectNotFoundException;
-import com.change_vision.jude.api.inf.project.ProjectAccessor;
 import com.change_vision.jude.api.inf.ui.IPluginActionDelegate;
 import com.change_vision.jude.api.inf.ui.IWindow;
 import hu.modeldriven.astah.core.AstahRepresentation;
@@ -13,9 +10,7 @@ import hu.modeldriven.core.eventbus.EventBus;
 
 import javax.swing.*;
 
-import static javax.swing.WindowConstants.EXIT_ON_CLOSE;
-
-public class TemplateAction implements IPluginActionDelegate {
+public class TraceabilityAction implements IPluginActionDelegate {
 
     public Object run(IWindow window) throws UnExpectedException {
         JFrame frame = new JFrame();
