@@ -5,20 +5,26 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 @SuppressWarnings({"squid:S3740"})
 public class EventBus {
 
     private final Map<Class<? extends Event>, Set<EventHandler<? extends Event>>> subscriptions = new ConcurrentHashMap<>();
 
-    @SuppressWarnings("unchecked")
+    private final ExecutorService executorService = Executors.newSingleThreadExecutor();
+
     public void publish(Event event) {
+        executorService.submit(() -> handleEvent(event));
+    }
+
+    private void handleEvent(Event event) {
         for (Map.Entry<Class<? extends Event>, Set<EventHandler<? extends Event>>> entry : subscriptions.entrySet()) {
-            if (entry.getKey().equals(event.getClass())) {
+            if (entry.getKey().isInstance(event)) {
                 for (EventHandler handler : entry.getValue()) {
                     handler.handleEvent(event);
                 }
-                return;
             }
         }
     }
