@@ -1,8 +1,8 @@
 package hu.modeldriven.astah.api;
 
 import com.change_vision.jude.api.inf.model.IPackage;
-import hu.modeldriven.astah.core.AstahException;
 import hu.modeldriven.astah.core.AstahRepresentation;
+import hu.modeldriven.astah.core.AstahRuntimeException;
 
 public class DummyAstahRepresentation implements AstahRepresentation {
 
@@ -13,6 +13,17 @@ public class DummyAstahRepresentation implements AstahRepresentation {
 
         DummyClass animal = new DummyClass("Animal");
         DummyClass dog = new DummyClass("Dog");
+
+        DummyClass myType = new DummyClass("MyType");
+        animal.addAttribute(new DummyAttribute(animal, "animalType", myType));
+        dog.addAttribute(new DummyAttribute(dog, "dogType", myType));
+
+        rootPackage.addElement(animal);
+        rootPackage.addElement(dog);
+
+        animal.addClientDependency(new DummyDependency("dep", animal, dog));
+
+        /*
         DummyClass element3 = new DummyClass("Element3");
         DummyClass element4 = new DummyClass("Element4");
         DummyClass element5 = new DummyClass("Element5");
@@ -46,14 +57,15 @@ public class DummyAstahRepresentation implements AstahRepresentation {
         childPackage.addElement(subChildPackage);
 
         subChildPackage.addElement(element5);
+        */
 
         return rootPackage;
     }
 
     @Override
-    public void selectModelElement(String id) throws AstahException {
+    public void selectModelElement(String id) {
         System.out.println("Selecting model element id=" + id);
-        throw new AstahException(new UnsupportedOperationException("Selecting model element"));
+        throw new AstahRuntimeException(new UnsupportedOperationException("Selecting model element"));
     }
 
 }
